+++ CHANGELOG.md
@@ -3,5 +3,11 @@
 #### Blocks:
 Adjusted how Heavy Air disables flight to reduce problematic conflict with other mod's creative flight.
 
+#### Fluids:
+Fixed bug where Sugar Water flowing down as a waterfall next to Sugar Cane can grow Sugar Cane beyond a height of 5.
+
+Added `the_bumblezone:sugar_water/grows_plant_faster` block tag that allows for adding other crops for Sugar Water to age faster.
+ Will work on Wheat and such if you add them to this tag.
+
 #### Configs:
 (Fabric): Fixed missing translations for some configs